+++ sound-mind-studio/src/main_window.cpp
@@ -205,6 +205,7 @@ MainWindow::MainWindow(QWidget* parent, sound_mind::core::AudioDeviceMode audioD
     connect(layersPanel_, &LayersPanel::renameRequested, this, &MainWindow::renameLayer);
     connect(layersPanel_, &LayersPanel::deleteRequested, this, &MainWindow::deleteLayer);
     connect(layersPanel_, &LayersPanel::duplicateRequested, this, &MainWindow::duplicateLayer);
+    connect(layersPanel_, &LayersPanel::cleanUpPhaseRequested, this, &MainWindow::cleanUpLayerPhase);
     connect(layersPanel_, &LayersPanel::reorderRequested, this, &MainWindow::reorderLayers);
     connect(layersPanel_, &LayersPanel::addLayerRequested, this, &MainWindow::addEmptyLayer);
     connect(layersPanel_, &LayersPanel::addFilterLayerRequested, this, &MainWindow::addFilterLayer);
@@ -1993,6 +1994,8 @@ bool MainWindow::renameLayerTo(sound_mind::core::LayerId id, const QString& newN
 void MainWindow::deleteLayer(sound_mind::core::LayerId id) { layerController_->deleteLayer(id); }
 
 void MainWindow::duplicateLayer(sound_mind::core::LayerId id) { layerController_->duplicateLayer(id); }
+
+void MainWindow::cleanUpLayerPhase(sound_mind::core::LayerId id) { layerController_->cleanUpLayerPhase(id); }
 
 void MainWindow::addEmptyLayer() {
     if (!project_) {